palm_diagnosis/requirements.txt
# ==============================
# Core Framework
# ==============================
Django>=4.2,<5.0

# ==============================
# Machine Learning & Deep Learning
# ==============================
tensorflow>=2.15
tf-keras>=2.15
numpy>=1.23

# ==============================
# Image Processing
# ==============================
Pillow>=9.5

# ==============================
# NLP & Embeddings (RAG)
# ==============================
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# ==============================
# Translation
# ==============================
deep-translator>=1.11.4

# ==============================
# Gemini API (Google GenAI)
# ==============================
google-genai>=0.3.0

# ==============================
# Utilities
# ==============================
python-dotenv>=1.0.0
